+++ controller/pagecontroller.php
@@ -160,6 +160,7 @@ class PageController extends Controller {
         $userFolder = \OC::$server->getUserFolder();
         $userfolder_path = $userFolder->getPath();
         $gpxcomp_root_url = "gpxvcomp";
+        $gpxedit_version = $this->config->getAppValue('gpxedit', 'installed_version');
 
         $this->cleanDbFromAbsentFiles(null);
 
@@ -219,6 +220,7 @@ class PageController extends Controller {
             'publicdir'=>'',
             'pictures'=>'',
             'token'=>'',
+            'gpxedit_version'=>$gpxedit_version,
             'gpxpod_version'=>$this->appVersion
         ];
         $response = new TemplateResponse('gpxpod', 'main', $params);